Separate classes and constants into individual files.
